+++ src/main/java/com/julian/codechallenge/calculator/controller/CalculatorController.java
@@ -1,4 +1,4 @@
-package com.julian.codechallenge.regexcalculator.controller;
+package com.julian.codechallenge.calculator.controller;
 
 import java.time.Duration;
 
@@ -10,8 +10,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.julian.codechallenge.calculator.service.CalculatorService;
+import com.julian.codechallenge.commons.UtilsBase64;
 import com.julian.codechallenge.commons.dto.Response;
-import com.julian.codechallenge.regexcalculator.service.RegexCalculatorService;
 
 import io.github.bucket4j.Bandwidth;
 import io.github.bucket4j.Bucket;
@@ -19,28 +20,23 @@ import io.github.bucket4j.Bucket4j;
 import io.github.bucket4j.Refill;
 
 @RestController
-@RequestMapping("/regex-calculator")
-public class RegexCalculatorController {
+@RequestMapping("/calculus")
+public class CalculatorController {
 
     @Autowired
-    private RegexCalculatorService regexCalculatorService;
+    private CalculatorService regexCalculatorService;
 
     private final Bucket bucket;
 
-    public RegexCalculatorController() {
+    public CalculatorController() {
         Bandwidth limit = Bandwidth.classic(20, Refill.greedy(20, Duration.ofMinutes(1)));
         this.bucket = Bucket4j.builder()
             .addLimit(limit)
             .build();
     }
 
-    @GetMapping("/hello")
-    public ResponseEntity<String> calculateEncodedQuery() {
-        return ResponseEntity.ok("Hello World!");
-    }
-    
-    @GetMapping("/calculus")
-    public ResponseEntity<Response> calculateEncodedQuery(@RequestParam("query") byte[] encodedQuery) {
+    @GetMapping
+    public ResponseEntity<Response> calculateEncodedQuery(@RequestParam("query") String encodedQuery) {
         if (bucket.tryConsume(1)) {
             return ResponseEntity.ok(regexCalculatorService.calculateEncodedQuery(encodedQuery));
         }
@@ -49,9 +45,9 @@ public class RegexCalculatorController {
     }
 
     @GetMapping("/encode")
-    public ResponseEntity<byte[]> encodeQuery(@RequestParam("query") String query) {
+    public ResponseEntity<String> encodeQuery(@RequestParam("query") String query) {
         if (bucket.tryConsume(1)) {
-            return ResponseEntity.ok(regexCalculatorService.encodeQuery(query));
+            return ResponseEntity.ok(UtilsBase64.encodeQuery(query));
         }
     
         return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
