+++ src/main/java/com/julian/codechallenge/calculator/service/CalculatorService.java
@@ -1,0 +1,9 @@
+package com.julian.codechallenge.calculator.service;
+
+import com.julian.codechallenge.commons.dto.Response;
+
+public interface CalculatorService {
+
+    public Response calculateEncodedQuery(String encodedQuery);
+
+}
