+++ src/main/java/com/julian/codechallenge/commons/UtilsBase64.java
@@ -1,0 +1,24 @@
+package com.julian.codechallenge.commons;
+
+import java.util.Base64;
+import java.util.Optional;
+
+public class UtilsBase64 {
+
+    private UtilsBase64() {}
+    
+    public static String encodeQuery(String query) {
+        return Base64.getEncoder().encodeToString(query.getBytes());
+    }
+
+	public static Optional<String> decodeQuery(String encodedQuery) {
+		try {
+			byte[] decodedBytes = Base64.getDecoder().decode(encodedQuery);
+			String decodedQuery = new String(decodedBytes);
+			return Optional.of(decodedQuery);
+		} catch (IllegalArgumentException exception) {
+			return Optional.empty();
+		}
+	}
+
+}
