--- src/main/java/com/julian/codechallenge/regexcalculator/service/RegexCalculatorService.java
@@ -1,11 +1,0 @@
-package com.julian.codechallenge.regexcalculator.service;
-
-import com.julian.codechallenge.commons.dto.Response;
-
-public interface RegexCalculatorService {
-
-    public Response calculateEncodedQuery(byte[] encodedQuery);
-
-    public byte[] encodeQuery(String query);
-
-}
